Added sample application to create a stand-alone executable for running/testing the kinetic producer and listener.  Also made enhancements to unit tests.
